remove audit policy validation from botanist
instead it is now validated inside the gardener-admission-controller
in a webhook
